% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{encode_base}
\alias{encode_base}
\title{One-hot encodes a DNA base}
\usage{
encode_base(base)
}
\arguments{
\item{base}{A,T,C,G or N}
}
\value{
A one-hot encoded vector
}
\description{
One-hot encodes a DNA base
}
